requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

torch==2.1.2
torchvision==0.16.2

diffusers==0.27.2
transformers==4.38.2
accelerate==0.27.2
safetensors==0.4.2

pillow==10.2.0
numpy<2.0
python-multipart==0.0.9
